fix responsibility
